requirements.txt
django==4.2.23
sqlparse==0.5.3
dotenv==0.9.9
psycopg2-binary==2.9.10
djangorestframework==3.16.1
django-cors-headers==4.7.0